// src/Interface/Interface.ts
export interface RegistrationInfo {
    studentName: string;
    studentAge: string;
    studentPhone: string;
    studentSchool: string;
    studentCourse: string;
    parentName: string;
    parentPhone: string;
    parentEmail: string;
    guardianName: string;
    guardianPhone: string;
    guardianEmail: string;
    yourPhone: string;
    
  }
  